requirements.txt
lancedb==0.6.13
sentence-transformers==2.7.0
pymupdf==1.24.3
python-docx==1.1.2
markdown==3.6
langchain-text-splitters==0.2.2
tiktoken==0.7.0
requests==2.32.3
python-dotenv==1.0.1
numpy==1.26.4
torch==2.3.1
pandas==2.2.2